Fix ColdLoad toast issues:
It was passing a string to a method which expected a transaction, so the method always returned failure. Detecting "success" by looking for "txid" in the text of the error message isn't reliable, because that string could also appear in a real error.

Changed to reuse the existing result checking code in Send.kt.